feat: 상품 단일 구매 서비스 추가 (#101)

+++ app/src/test/java/shop/woowasap/mock/dto/OrderProductRequest.java
@@ -1,4 +1,4 @@
 package shop.woowasap.mock.dto;
 
-public record OrderProductRequest(int quantity) {
+public record OrderProductRequest(long quantity) {
 }
