+++ order/order-domain/src/main/java/shop/woowasap/order/domain/exception/InvalidQuantityException.java
@@ -4,7 +4,7 @@ import java.text.MessageFormat;
 
 public class InvalidQuantityException extends RuntimeException {
 
-    public InvalidQuantityException(int quantity) {
+    public InvalidQuantityException(long quantity) {
         super(MessageFormat.format("quantity \"{0}\" 는 0 이하가 될 수 없습니다.", quantity));
     }
 }
